requirements.txt
aiofile
anyio
async-timeout
caio
certifi
charset-normalizer
click
cmake
fastapi
ffmpeg-python
filelock
future
h11
idna
importlib-metadata
Jinja2
lit
llvmlite
MarkupSafe
more-itertools
mpmath
networkx
numba
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
openai-whisper
pkg_resources
pydantic
python-multipart
pyzmq
regex
requests
sniffio
starlette
sympy
tiktoken
torch
tqdm
triton
typing_extensions
urllib3
uvicorn
zipp
